fix:mobile menu in react
